refactor: 메인페이지 코드 리펙토링
